requirements.txt
playwright
flask
pandas
python-dotenv
Flask-CORS 